+++ spec/document/elements/table/properties/table_caption_spec.rb
@@ -1,0 +1,8 @@
+require 'spec_helper'
+
+describe 'table_caption' do
+  it 'table_caption' do
+    docx = OoxmlParser::Parser.parse('spec/document/elements/table/properties/caption/table_caption.docx')
+    expect(docx.elements[1].properties.caption.value).to eq('Simple Test Text')
+  end
+end