+++ spec/document/elements/table/properties/table_description_spec.rb
@@ -1,0 +1,8 @@
+require 'spec_helper'
+
+describe 'table_caption' do
+  it 'table_caption' do
+    docx = OoxmlParser::Parser.parse('spec/document/elements/table/properties/description/table_description.docx')
+    expect(docx.elements[1].properties.description.value).to eq('SimpleTestText')
+  end
+end